requirements.txt
# Core workflow
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.2.0
langchain-tavily>=0.1.0
langchain-experimental>=0.0.60
langchain-google-genai>=1.0.0
langchain-deepseek>=0.1.0
openai>=1.0.0

# Utilities
pydantic>=2.0.0
jinja2>=3.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
requests>=2.31.0
httpx>=0.27.0
json-repair>=0.30.0
markdownify>=0.12.0
readabilipy>=0.2.0
aiohttp>=3.9.0

# Optional RAG / Vector DB (enable if you use these providers)
qdrant-client>=1.9.0
langchain-qdrant>=0.1.0
pymilvus>=2.4.0
langchain-milvus>=0.1.0
